Added test interface, DLL's and a test driver
I have added the interface and 3 more DLL's one of which is the test driver named TestInterface.  From here the other DLL test lib files are loaded and run.  Also I added a final XML file that is parsed to retrieve the DLL files.  I did make sure to change the build order that all the DLL's and Test Interface are build before the test harness.